# Defer windowWidth() to improve perf for non-help usage