--- a/lib/usage.js
+++ b/lib/usage.js
@@ -96,9 +96,21 @@ module.exports = function (yargs, y18n) {
     epilog = msg
   }
 
-  var wrap = windowWidth()
+  var wrapSet
+  var wrap
   self.wrap = function (cols) {
+    wrapSet = true
     wrap = cols
+  }
+
+  function getWrap () {
+    // lazily call windowWidth() because it's very expensive,
+    // and only needs to be called if the user wants to show usage/help
+    if (!wrapSet) {
+      wrap = windowWidth()
+    }
+    wrapSet = true
+    return wrap
   }
 
   var deferY18nLookupPrefix = '__yargsString__:'
@@ -122,9 +134,10 @@ module.exports = function (yargs, y18n) {
         return acc
       }, {})
     )
+    var theWrap = getWrap()
     var ui = require('cliui')({
-      width: wrap,
-      wrap: !!wrap
+      width: theWrap,
+      wrap: !!theWrap
     })
 
     // the usage string.
@@ -271,7 +284,8 @@ module.exports = function (yargs, y18n) {
 
     // if we've enabled 'wrap' we should limit
     // the max-width of the left-column.
-    if (wrap) width = Math.min(width, parseInt(wrap * 0.5, 10))
+    var theWrap = getWrap()
+    if (theWrap) width = Math.min(width, parseInt(theWrap * 0.5, 10))
 
     return width
   }