--- a/lib/usage.js
+++ b/lib/usage.js
@@ -96,7 +96,7 @@ module.exports = function (yargs, y18n) {
     epilog = msg
   }
 
-  var wrapSet
+  var wrapSet = false
   var wrap
   self.wrap = function (cols) {
     wrapSet = true
@@ -108,8 +108,8 @@ module.exports = function (yargs, y18n) {
     // and only needs to be called if the user wants to show usage/help
     if (!wrapSet) {
       wrap = windowWidth()
+      wrapSet = true
     }
-    wrapSet = true
     return wrap
   }