--- a/lib/usage.js
+++ b/lib/usage.js
@@ -153,7 +153,7 @@ module.exports = function (yargs, y18n) {
 
       commands.forEach(function (command) {
         ui.div(
-          {text: command[0], padding: [0, 2, 0, 2], width: maxWidth(commands) + 4},
+          {text: command[0], padding: [0, 2, 0, 2], width: maxWidth(commands, theWrap) + 4},
           {text: command[1]}
         )
       })
@@ -227,7 +227,7 @@ module.exports = function (yargs, y18n) {
         ].filter(Boolean).join(' ')
 
         ui.span(
-          {text: kswitch, padding: [0, 2, 0, 2], width: maxWidth(switches) + 4},
+          {text: kswitch, padding: [0, 2, 0, 2], width: maxWidth(switches, theWrap) + 4},
           desc
         )
 
@@ -248,7 +248,7 @@ module.exports = function (yargs, y18n) {
 
       examples.forEach(function (example) {
         ui.div(
-          {text: example[0], padding: [0, 2, 0, 2], width: maxWidth(examples) + 4},
+          {text: example[0], padding: [0, 2, 0, 2], width: maxWidth(examples, theWrap) + 4},
           example[1]
         )
       })
@@ -267,7 +267,7 @@ module.exports = function (yargs, y18n) {
 
   // return the maximum width of a string
   // in the left-hand column of a table.
-  function maxWidth (table) {
+  function maxWidth (table, theWrap) {
     var width = 0
 
     // table might be of the form [leftColumn],
@@ -284,7 +284,6 @@ module.exports = function (yargs, y18n) {
 
     // if we've enabled 'wrap' we should limit
     // the max-width of the left-column.
-    var theWrap = getWrap()
     if (theWrap) width = Math.min(width, parseInt(theWrap * 0.5, 10))
 
     return width