Hi :wave:, I'm a big fan of yargs, so after talking with @lrlna about https://github.com/yargs/yargs/issues/521, I wanted to try my hand at improving the overall runtime of `require('yargs')`.

I started by profiling with the Chrome dev tools using `node-nightly` and the `--inspect` option, where I noticed that `windowWidth()` is a very expensive function:

![screenshot 2016-08-29 09 12 12](https://cloud.githubusercontent.com/assets/283842/18064626/10d66122-6de5-11e6-9564-2ae9dcbdd5b1.png)


Since `windowWidth()` only has to be calculated when we need the column width, e.g. when the user has typed `--help` or `-h`, I figured we could lazily evaluate it. That's what this PR does.

Here is a very small benchmark to demonstrate the perf improvement:

```bash
time for i in {1..500}; do node -e 'require("./index")'; done
```

Running in node 5.12.0 on a 2013 Macbook Air, I got the following results (real time, averaged):

* **Before:** 122.14ms
* **After:** 108.746ms

Since that includes the cost of the Node process itself, we can also use `performance-now` to measure just the runtime of `require('yargs')`:

```bash
for ((i=0;i<500;i++)); do node -e 'var now = require("performance-now"); var s = now(); require("./index"); var e = now() - s; console.log(e)'; done | awk '{ sum+=$1 } END { print sum/NR }'
```

This gives: 

* **Before:** 41.3002ms
* **After:** 28.5307ms

So in rough terms `require('yargs')` should be about ~30% faster with this PR, with an average of ~12ms shaved off each invocation of `require('yargs')`. (Obviously it doesn't make any difference when doing `-h` or `--help`, but my hunch is that that's not the average use case.)

Please let me know if I missed something in the style/coverage/etc. or if this PR is unfeasible for some other reason I didn't notice. Hope this helps!