@nexdrew 2016-08-29
Besides my nit-picky inline comments, this LGTM! Thanks @nolanlawson!