@bcoe 2016-08-29
@nolanlawson thank you very much for digging into this.