@bcoe 2016-09-30
@nolanlawson @lrlna want to give this a spin?

`npm cache clear; yargs@6.0.0-alpha.1`?

These performance changes go out with the next release (would love the extra set of eyes).